Added new exception and deletion functionality
Added a new exception for when the system tries to access a user that
does not exist.

Also added new functionality for User and RoomScheduler classes that
allows a RoomScheduler to add/remove a reservation that is stored in the
user's account at the same time it is added/removed from the scheduler.